+++ test/enum.ts
@@ -1,6 +1,5 @@
 import { GraphQLEnumType } from 'graphql';
-import { getGraphQLType } from '../src/types/reflection';
-import { EnumType } from '../src/types/index';
+import { getGraphQLType, EnumType } from '../src';
 
 describe('Enum creation', () => {
   test(`create Enum`, () => {
@@ -68,44 +67,3 @@ describe('Enum creation', () => {
     ]);
   });
 });
-
-// describe('Enum can be mounted', () => {
-//   class MyEnum extends Enum {
-//     static description = "Desc";
-//     static values = {
-//       VALUE1: {
-//         description: "Value1 description",
-//         value: 1
-//       },
-//       VALUE2: {
-//         description: "Value2 description",
-//         deprecationReason: "Value2 is deprecated",
-//         value: 2
-//       }
-//     }
-//   }
-
-//   test(`as a Field`, () => {
-//     let unmounted = new Enum({description: "MyEnum field"});
-//     let field = unmounted.toField();
-//     expect(field).toBeInstanceOf(Field);
-//     expect(field.type).toBe(Enum);
-//     expect(field.options).toMatchObject({description: "MyEnum field"});
-//   })
-
-//   test(`as a Argument`, () => {
-//     let unmounted = new Enum({description: "MyEnum argument"});
-//     let arg = unmounted.toArgument();
-//     expect(arg).toBeInstanceOf(Argument);
-//     expect(arg.type).toBe(Enum);
-//     expect(arg.options).toMatchObject({description: "MyEnum argument"});
-//   })
-
-//   test(`as a InputField`, () => {
-//     let unmounted = new Enum({description: "MyEnum input field"});
-//     let inputfield = unmounted.toInputField();
-//     expect(inputfield).toBeInstanceOf(InputField);
-//     expect(inputfield.type).toBe(Enum);
-//     expect(inputfield.options).toMatchObject({description: "MyEnum input field"});
-//   })
-// })
