+++ test/structures.ts
@@ -1,6 +1,5 @@
 import { GraphQLList, GraphQLString, GraphQLNonNull } from 'graphql';
-import { List, NonNull } from './../src/types';
-import { getGraphQLType } from '../src/types/reflection';
+import { getGraphQLType, List, NonNull } from './../src';
 // import { GraphQLID, GraphQLList, GraphQLNonNull } from 'graphql';
 // import { Field, InputField, Argument } from '../src';
 // import { ID } from './../src/types/scalars';
